+++ apps/triggers/src/app.module.ts
@@ -23,6 +23,7 @@ import { StatsModule } from './stats/stat.module';
 import { HttpModule } from '@nestjs/axios';
 import { SettingsModule } from '@lib/core';
 import { LibraryModule } from './library/library.module';
+import { SseModule } from './sse/sse.module';
 
 @Module({
   imports: [
@@ -77,6 +78,7 @@ import { LibraryModule } from './library/library.module';
     TriggerHistoryModule,
     StatsModule,
     LibraryModule,
+    SseModule,
   ],
   controllers: [AppController],
   providers: [AppService],
