+++ apps/triggers/src/phases/phases.module.ts
@@ -1,12 +1,12 @@
 import { BullModule } from '@nestjs/bull';
 import { forwardRef, Module } from '@nestjs/common';
-import { BQUEUE, MS_TRIGGER_CLIENTS, SSE_EVENTS } from 'src/constant';
+import { BQUEUE, MS_TRIGGER_CLIENTS } from 'src/constant';
 import { PhasesController } from './phases.controller';
 import { PhasesService } from './phases.service';
 import { ClientsModule, Transport } from '@nestjs/microservices';
 import { TriggerModule } from 'src/trigger/trigger.module';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import Redis from 'ioredis';
+import { SseModule } from 'src/sse/sse.module';
 
 @Module({
   imports: [
@@ -37,19 +37,7 @@ import Redis from 'ioredis';
     ]),
   ],
   controllers: [PhasesController],
-  providers: [
-    PhasesService,
-    {
-      provide: SSE_EVENTS.PUBLISHER,
-      useFactory: (config: ConfigService) =>
-        new Redis({
-          host: config.get('REDIS_HOST'),
-          port: Number(config.get('REDIS_PORT')),
-          password: config.get('REDIS_PASSWORD'),
-        }),
-      inject: [ConfigService],
-    },
-  ],
+  providers: [PhasesService, SseModule],
   exports: [PhasesService],
 })
 export class PhasesModule {}
