+++ apps/triggers/src/processors/processors.module.ts
@@ -7,10 +7,9 @@ import { CommunicationProcessor } from './communication.processor';
 import { ActivityModule } from 'src/activity/activity.module';
 import { StatsModule } from 'src/stats/stat.module';
 import { ClientsModule, Transport } from '@nestjs/microservices';
-import { CORE_MODULE, SSE_EVENTS } from 'src/constant';
+import { CORE_MODULE } from 'src/constant';
 import { NotificationProcessor } from './notification.processor';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import Redis from 'ioredis';
 
 @Module({
   imports: [
@@ -39,16 +38,6 @@ import Redis from 'ioredis';
     CommunicationProcessor,
     StatsProcessor,
     NotificationProcessor,
-    {
-      provide: SSE_EVENTS.PUBLISHER,
-      useFactory: (config: ConfigService) =>
-        new Redis({
-          host: config.get('REDIS_HOST'),
-          port: Number(config.get('REDIS_PORT')),
-          password: config.get('REDIS_PASSWORD'),
-        }),
-      inject: [ConfigService],
-    },
   ],
 })
 export class ProcessorsModule {}
