+++ apps/triggers/src/sse/sse.module.ts
@@ -1,0 +1,24 @@
+import { Global, Module } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import Redis from 'ioredis';
+import { SSE_EVENTS } from 'src/constant';
+import { SseService } from './sse.service';
+
+@Global()
+@Module({
+  providers: [
+    {
+      provide: SSE_EVENTS.PUBLISHER,
+      useFactory: (config: ConfigService) =>
+        new Redis({
+          host: config.get('REDIS_HOST'),
+          port: Number(config.get('REDIS_PORT')),
+          password: config.get('REDIS_PASSWORD'),
+        }),
+      inject: [ConfigService],
+    },
+    SseService,
+  ],
+  exports: [SseService],
+})
+export class SseModule {}
