+++ apps/triggers/src/trigger/trigger.module.ts
@@ -2,13 +2,12 @@ import { forwardRef, Module } from '@nestjs/common';
 import { TriggerService } from './trigger.service';
 import { TriggerController } from './trigger.controller';
 import { BullModule } from '@nestjs/bull';
-import { BQUEUE, CORE_MODULE, SSE_EVENTS } from 'src/constant';
+import { BQUEUE, CORE_MODULE } from 'src/constant';
 import { PhasesModule } from 'src/phases/phases.module';
 import { HttpModule } from '@nestjs/axios';
 import { SourcesDataModule } from 'src/sources-data/sources-data.module';
 import { ClientsModule, Transport } from '@nestjs/microservices';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import Redis from 'ioredis';
 
 @Module({
   imports: [
@@ -40,19 +39,7 @@ import Redis from 'ioredis';
     forwardRef(() => PhasesModule),
   ],
   controllers: [TriggerController],
-  providers: [
-    TriggerService,
-    {
-      provide: SSE_EVENTS.PUBLISHER,
-      useFactory: (config: ConfigService) =>
-        new Redis({
-          host: config.get('REDIS_HOST'),
-          port: Number(config.get('REDIS_PORT')),
-          password: config.get('REDIS_PASSWORD'),
-        }),
-      inject: [ConfigService],
-    },
-  ],
+  providers: [TriggerService],
   exports: [TriggerService],
 })
 export class TriggerModule {}
